addressed Katherine's feedback
