+++ jakarta.jdt/org.eclipse.lsp4jakarta.jdt.core/src/main/java/org/eclipse/lsp4jakarta/jdt/core/websocket/AddPathParamQuickFix.java
@@ -25,12 +25,4 @@ public class AddPathParamQuickFix extends InsertAnnotationQuickFix {
     public AddPathParamQuickFix() {
         super("jakarta.websocket.server.PathParam", false, "value");
     }
-
-    @Override
-    protected String getLabel(String annotation, String... attributes) {
-        StringBuilder name = new StringBuilder("Insert ");
-        name.append("@");
-        name.append(annotation);
-        return name.toString();
-    }
 }
